update: Move db functions to their own files.
