+++ commands/category/remove.ts
@@ -2,11 +2,12 @@ import { CommandInteraction, Permissions } from 'discord.js-light';
 import { Category } from '../../utilities/types/commands';
 import { SlashCommand } from '../slashCommand';
 import RoleBot from '../../src/bot';
+
+import { handleInteractionReply } from '../../utilities/utils';
 import {
   DELETE_CATEGORY_BY_ID,
   GET_CATEGORY_BY_NAME,
-} from '../../src/database/database';
-import { handleInteractionReply } from '../../utilities/utils';
+} from '../../src/database/queries/category.query';
 
 export class RemoveCategoryCommand extends SlashCommand {
   constructor(client: RoleBot) {
@@ -40,7 +41,11 @@ export class RemoveCategoryCommand extends SlashCommand {
         `Required option was empty for categoryName[${categoryName}]`,
         interaction.guildId
       );
-      return handleInteractionReply(this.log, interaction, `Hey! I don't think you passed in a name. Could you please try again?`);
+      return handleInteractionReply(
+        this.log,
+        interaction,
+        `Hey! I don't think you passed in a name. Could you please try again?`
+      );
     }
 
     const category = await GET_CATEGORY_BY_NAME(
@@ -54,7 +59,11 @@ export class RemoveCategoryCommand extends SlashCommand {
         interaction.guildId
       );
 
-      return handleInteractionReply(this.log, interaction, `Hey! I could **not** find a category by the name of \`${categoryName}\`. This command is case sensitive to ensure you delete exactly what you want. Check the name and try again.`);
+      return handleInteractionReply(
+        this.log,
+        interaction,
+        `Hey! I could **not** find a category by the name of \`${categoryName}\`. This command is case sensitive to ensure you delete exactly what you want. Check the name and try again.`
+      );
     }
 
     DELETE_CATEGORY_BY_ID(category.id)
@@ -64,7 +73,11 @@ export class RemoveCategoryCommand extends SlashCommand {
           interaction.guildId
         );
 
-        handleInteractionReply(this.log, interaction, `Hey! I successfully deleted the category \`${categoryName}\` for you and freed all the roles on it.`);
+        handleInteractionReply(
+          this.log,
+          interaction,
+          `Hey! I successfully deleted the category \`${categoryName}\` for you and freed all the roles on it.`
+        );
       })
       .catch((e) => {
         this.log.error(
@@ -72,7 +85,11 @@ export class RemoveCategoryCommand extends SlashCommand {
           interaction.guildId
         );
 
-        handleInteractionReply(this.log, interaction, `Hey! I had an issue deleting the category. Please wait a second and try again.`);
+        handleInteractionReply(
+          this.log,
+          interaction,
+          `Hey! I had an issue deleting the category. Please wait a second and try again.`
+        );
       });
   };
 }
